requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
qdrant-client
anthropic
beautifulsoup4
atlassian-python-api